pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "woocommerce-mcp"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server for WooCommerce integration"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "mcp>=0.1.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "ruff>=0.0.267",
    "mypy>=1.3.0",
    "pytest>=7.3.1",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
woocommerce-mcp = "woocommerce_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/woocommerce_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*" 